requirements/local.txt
-r base.txt

django-extensions
ipdb
